Refactor zng-ext-fs-watcher code
